Fix sidebar drag-resize; split into primitives/ + layout/
Two coupled changes:

1. Restore the drag-to-resize / drag-to-collapse handle. The
   migration to Vue dropped the imperative
   `workspace.style.setProperty("--sidebar-w", …)` from the old
   pointer handler in favour of a reactive `sidebar.width` ref —
   but nothing was binding that ref back into CSS. The grid
   column kept its `:root` default of 280px no matter how the
   user dragged. `pages/layout.vue` now applies `--sidebar-w` to
   the workspace element via `:style`, so resizing, collapsing,
   and the resizer's `right: calc(var(--sidebar-w) - 3px)` all
   move together.

2. Reorganise by the layers from style-guide.vue:
   * `primitives/Sidebar.vue` — generic aside primitive. Holds
     positioning (left/right border, grid placement via the
     parent's `:deep(.sidebar)` rule, mobile bottom-sheet drawer
     translate). No business logic. Slots: `handle`, `head`,
     default body.
   * `layout/SidebarRight.vue` — composes the primitive with the
     workspace's actual content (the Explorer header, the close
     button for the mobile sheet, the file tree) and binds it to
     the `useSidebar` state.
   * Everything that used to live in `features/shell/` (TitleBar,
     SidebarTree, TreeNode, SidebarResizer, TabBar, EditorTab,
     BottomNav, plus the three composables) moves to `layout/`,
     matching the user's mental model that these are "layout
     pieces" rather than a domain feature.
