+++ apps/web/src/layout/SidebarResizer.vue
